// src/interfaces/patients.ts
export interface IPatienTypes {
  id?: string;
  gender?: string;
  status?: string;
  createdAt?: Date;
  updatedAt?: Date;
  firstName?: string;
  lastName?: string;
  dob?: Date;
  phone?: string;
  email?: string;
  address?: string;
  profile?: string;
  balance?: string;
  role?: string;
  banks?: [];
}